move maxUnstakeAmount and refactor handleAddDelegation
